requirements.txt
flask
slack_sdk
google-generativeai
gunicorn